requirements.txt
python-dotenv==0.19.2
python-telegram-bot==13.8.1
psycopg2-binary==2.8.6
phonenumbers==8.12.37
pyqrcode==1.2.1
pypng==0.0.21
